style: 코드 컨벤션에 따른 주석 제거 및 rootViewController 설정
