+++ Diary/Resources/Assets.xcassets/Contents.json
